backend/phase1_data_collection/requirements.txt
# Data Collection & Aggregation Dependencies

# Core async libraries
aiohttp>=3.8.0
aioredis>=2.0.0
asyncio-mqtt>=0.11.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
pydantic>=1.10.0

# Database connectors
psycopg2-binary>=2.9.0
pymongo>=4.0.0
redis>=4.0.0

# Cloud storage
boto3>=1.26.0
google-cloud-storage>=2.0.0
azure-storage-blob>=12.0.0

# Message queues
kafka-python>=2.0.0
google-cloud-pubsub>=2.0.0
azure-eventhub>=5.0.0

# Data validation
jsonschema>=4.0.0
cerberus>=1.3.0

# Logging and monitoring
structlog>=22.0.0
prometheus-client>=0.15.0

# Utilities
python-dateutil>=2.8.0
pytz>=2022.1
hashlib2>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.8.0

# Development
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
